Creates image based on tailscale source code (#1)
TL;DR
-----

Uses the tailscale source code to build an image for the proxy

Details
-------

Implements a simple pipeline that creates an image for the proxy
from the example in the Tailscale source code. The pipeline uses
the `Dockerfile` provided in the source repo to build the image,
the pushes it to the local registry.

Also implements a script `setup-harbor` the prepares the local
registry by configuring the `network` project that the new image
gets pushed to. There's also are replication setup so that the new
image can be build based on a local version of the base image that
Tailscale provides on GitHub.

The replication is currently unused since we leverage the existing
source code to build the image.